restructure code
Merge Main.java with App.java.
Rename "controller" package to "app"
Refactor RecordType enum out of SwingView to separate file in model package.
