+++ src/edu/umgc/smart/model/RecordType.java
@@ -1,0 +1,5 @@
+package edu.umgc.smart.model;
+
+public enum RecordType {
+  REPORT, LETTER, MEMO, PRESENTATION, TECHNICAL
+}
